Implement a basic heartbeat process
This creates two keys with TTLs:

- `node/{name}/healthy`, 30s TTL
- `node/{name}/alive`, 300s TTL

And starts two async processes to refresh the TTLs every 10s and 100s
respectively.  If connection to etcd is lost, or renewing the lease
fails, it attempts to reconnect to etcd and recreate the lease.

If reconnecting to etcd fails, it retries with exponential backoff.  If
it fails to reconnect for 10 attempts in a row, the process terminates.

The reason for the two keys with different TTLs is a node may be briefly
unavailable due to a network outage.  I think it's useful to know that a
node *may* have died (and so to avoid giving it new work), but to wait a
little longer before reallocating all work that it has.