Implement a basic worker process
To facilitate this, this also adds a basic gRPC API to the supervisor
process which the worker uses to register itself and also to issue
heartbeats.

Like the supervisor, the worker gets both `healthy` and `alive` keys and
TTLs, though the worker heartbeat refreshes both of those leases.

This commit also adds a new key namespace to etcd: `lease/{id}`, which
has the value `node/{name}/(alive|healthy)`.  So, if the `healthy` lease
has expired, the supervisor can work out the worker's name from the
`alive` lease, and so recreate the `healthy` lease.