Split out a nodelib crate and have workerd talk to etcd
I initially wanted all communication with etcd to go via supervisord.  I
felt it was probably better to have a relatively small number of nodes
which talk to etcd, rather than hundreds / thousands of workerd nodes.

But I've since learned that flannel, the container networking layer I
intend to use, has every node talk to etcd.  Well, the flannel people
have presumably done their research and decided it scales well enough,
so why shouldn't I trust them on that?

This also greatly simplifies the architecture: workerd doesn't need to
talk to supervisord at all, it can talk to etcd directly to heartbeat
and watch for new pods assigned to it; supervisord can just do
scheduling, by writing to the keys that workerd is watching (and so I've
now renamed it to schedulerd); there will be an eventual reaperd that
watches for dead workerd nodes and changes the state of its pods to
"dead"; etc.  Lots of little simple processes, each doing one thing and
doing it well.