Enforce node name uniqueness
A node cannot be started with a name if the liveness lease for that name
is still active.

This means that if a node crashes it either has to wait 5 minutes before
restarting for the lease to elapse, or use a different name - the next
commit will introduce a random name generator to address that.