Randomly generate a unique node name if unspecified
The format is `a{base36(ip)}-p{base36(port)}-{random}`.

The address and port part ensure uniqueness across the cluster, as two
running processes can't bind to the same address and port at the same
time.  The random suffx ensures that if a process crashes and gets
restarted, the new instance will (very likely) generate a different name
to before (and if it doesn't, it'll crash-loop until it does).