Fix watch deadlocks
The channels had a capacity of 16, which means that writing a 17th
element blocks until the 1st has been consumed.  But the reader and
writer processes are both taking out locks on the state, which means
that the reader can't process anything until the writer finishes.

So fix that by using `try_send`, which'll just abort if the channel is
full.  The elements will then be picked up by the retry task.

Also increase the channel capacity so there's less blocking anyway.