Implement a proof-of-concept pod-worker
This takes a command and set of environment variables from the pod spec,
runs it, and uses the exit code to set the status in the resource json
in etcd.  No containers yet, but this proves the concept: I can create
pods, schedulers assign those pods to workers, and workers run those
pods and report whether they were successful or not.