Run actual containers and pods
Well, this is it.  A real proof-of-concept distributed container
scheduler.  Schedulers allocate pods to workers based on resource
availability, workers create and execute containers under the given
resource constraints and report their status back.

Sure there's no DNS yet, or a way to get the logs of a pod, or even many
basic validations (like enforcing global pod name uniqueness, or
container name uniqueness within a pod, or ensuring two containers in a
pod don't try to expose the same port).  And I've not actually tested
that the networking works.  But this is a really good start!

Annoyingly, this is kind of slow.  I wonder if a channel is getting
backed up or something.