Implement cluster DNS
Along with some documentation on how to set up the networking side of
things, as I've just spent two days trying to figure this out.

Initially I wanted to implement DNS as a pod running in the cluster.
But I decided not to: while I *could* run a DNS service inside a pod,
it's kind of a special service that `workerd` needs to know about, as it
needs to be able to pass the nameserver IP into all the containers.
Tightly coupling an out-of-cluster process with an in-cluster process
didn't seem a great pattern, so I put DNS in `workerd`.

This only supports A records, and only supports a single IP per
hostname.  That's good enough for nodes and pods, but won't be good
enough for services, which'll be associated with multiple pods.  So I
think a next iteration of this will introduce something a little
smarter.  I'll also need to share the DNS record creation code with the
future `apid`, so some of this will likely move into the `nodelib`.