Implement DNS aliases
An alias is a name that refers to one or more other names, where the
records of those other names automatically propagate to the original
name.

For example, say we have two pods:

- `blasphemous-bumblebee-1234.pod.cluster.local.` at `10.5.88.3`
- `stygian-hedgehog-9876.pod.cluster.local.` at `10.5.66.7`

And these pods are interchangeable, so we'd like to have one domain name
that refers to the both of them.

We can set up the alias `megapod.svc.cluster.local.` by creating the
etcd keys:

- `/thing-doer/dns/megapod.svc.cluster.local./blasphemous-bumblebee-1234.pod.cluster.local.`
- `/thing-doer/dns/megapod.svc.cluster.local./stygian-hedgehog-9876.pod.cluster.local.`

And then when a query comes in for `megapod.svc.cluster.local.` the
result will be the combined answers for the other two names.

Aliases can refer to aliases, and results are deduplicated.

Right now, all worker nodes add themselves to
`dns.special.cluster.local.`.  The API server will be at
`api.special.cluster.local.`.