Implement a basic API server
This is a large commit, but the changes are fairly self-explanatory (and
a quarter of the diff is just Cargo.lock changes due to new dependencies
anyway).

I went for an HTTP API to start with, rather than my initial thinking of
a gRPC API, just to make it easier to integrate with.  A gRPC API might
come in the future as I look into supporting features like watches.

In the more immediate future though, I think I want to:

- Write a basic API client integration test tool
- Implement better error handling for invalid requests (see TODO)
- Document the API and its errors (see TODO)
- Document the special behaviour of the node and pod resource types