Bring back the reaperd retry tasks
By merging these into the processing tasks I introduced a bug - things
would only be retries if nothing new to process arrived in the
`RETRY_INTERVAL`, as processing a message repeats the loop and so resets
the timer.

This commit brings the tasks back, but keeps the logic of which things
to retry in the main task, so that state at least doesn't need to be
shared.