Add support for basic scheduling constraints
This commit allows a pod to specify a list of workers that it *may* be
scheduled on, and a list of workers that it *cannot* be scheduled on.
The workers which meet the resource requirements are filtered by these
two lists.

This is a really basic building block of ReplicaSets: if the desired
number of pods doesn't match the expected number, it'll create more pods
with the constraint that they're not scheduled to any of the workers
currently running one of the ReplicaSets' pods.  A more robust approach
would be being able to say something like "don't schedule to a worker
that is running a pod with such-and-such metadata", but right now the
scheduler doesn't directly know anything about the pods a worker is
running, and I want to think a bit more about the best way to do that.