Fix pod port mapping
This commit properly implements port mapping by adding iptables
rules to the `nat` `PREROUTING` (for external connections) and
`OUTPUT` (for internal connections) chains.  The new rules are all put
into a dedicated chain named after the node (by default - there's a
command line argument to rename it).

Since different chains are invoked for internal and external
connections, I've changed the integration test to start curl pods on
both nginx nodes.

I've also moved the `ContainerPortSpec` into the `PodSpec` directly,
since it's really a property of the pod and not the containers inside
the pod.