Implement container firewall
This commit implements the container firewall by adding iptables rules
to the `filter` `FORWARDING` (for container to container, or external
host to container) and `OUTPUT` (for host to container) chains.  The new
rules are put into a dedicated chain with the same name as the `nat`
rules (though these are different chains, they just have the same name).

In testing this I discovered a race condition where a schedulerd may
miss a workerd if they start up concurrently.  I've adjusted the
integration test to avoid the bug and will open a separate issue to fix
it.